# inicial/Clientes/Controllers/AdminsApi.cs
/*
 * Clientes API
 *
 * API de Clientes
 *
 * OpenAPI spec version: 1.0.0
 * Contact: michael.costa@localiza.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;

using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class AdminsApiController : ControllerBase
    { 
        /// <summary>
        /// Adiciona um cliente
        /// </summary>
        /// <remarks>adiciona um cliente</remarks>
        /// <param name="cliente">Cliente a ser inserido</param>
        /// <response code="201">Cliente criado</response>
        /// <response code="400">entrada inválida, objeto inválido</response>
        /// <response code="409">Um cliente com esses dados já existe</response>
        [HttpPost]
        [Route("/michaelcosta/Cliente/1.0.0/clientes")]
        [ValidateModelState]
        [SwaggerOperation("AdicionarCliente")]
        public virtual IActionResult AdicionarCliente([FromBody]Cliente cliente)
        { 
            //TODO: Uncomment the next line to return response 201 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(201);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 409 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(409);


            throw new NotImplementedException();
        }
    }
}